# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "fluent liquidity mapper v1.0",
        "contact": {
            "name": "Pavlo Hrishyn",
            "email": "pashagrishyn@gmail.com"
        },
        "license": {}
    },
    "host": "localhost:4421",
    "basePath": "/api",
    "paths": {
...
        "/liquidity/:provider": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "saves in database new record and sets it in cache",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Liquidity"
                ],
                "summary": "creates new liquidity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "provider",
                        "name": ":provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "liquidity",
                        "name": "liquidity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/liquidity.Liquidance"
                        }
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
...
    },
    "definitions": {
        "liquidity.Liquidance": {
            "type": "object",
            "required": [
                "btc",
                "eur",
                "usd"
            ],
            "properties": {
                "btc": {
                    "type": "number"
                },
                "eur": {
                    "type": "number"
                },
                "usd": {
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-Authorization",
            "in": "header"
        }
    }
}